refactor(peagen): split orm tables into modules and fix tenant bootstrap
