+++ pkgs/standards/peagen/peagen/orm/__init__.py
@@ -1,54 +1,33 @@
 """
-peagen/orm/__init__.py  –  all Peagen domain tables in one place
-(uses only sqlalchemy.Column – no mapped_column / JSONB).
+peagen.orm – aggregate and expose ORM table classes.
 """
 
 from __future__ import annotations
 
 from typing import FrozenSet
 
-from autoapi.v2.types import (
-    Column,
-    Integer,
-    String,
-    UniqueConstraint,
-    relationship,
-)
+from autoapi.v2.tables import Org, Role, RoleGrant, RolePerm, Status, Base
 
-# ---------------------------------------------------------------------
-# bring in the baseline tables that AutoAPI already owns
-# ---------------------------------------------------------------------
-from autoapi.v2.tables import Tenant as TenantBase, User as UserBase, Org
-from autoapi.v2.tables import Role, RoleGrant, RolePerm
-from autoapi.v2.tables import Status
-from autoapi.v2.tables import Base
-from autoapi.v2.mixins import (
-    GUIDPk,
-    # TenantMixin,
-    UserMixin,
-    Ownable,
-    Bootstrappable,
-    Timestamped,
-    TenantBound,
-    StatusMixin,
-    BlobRef,
-)
-
-from .pools import Pool
-from .workers import Worker
+# Import table classes. Ensure Tenant is imported before Pool so bootstrapping
+# default rows inserts the default tenant prior to default pools.
+from .tenants import Tenant
+from .users import User
+from .repositories import Repository
+from .user_repositories import UserRepository
+from .raw_blobs import RawBlob
 from .keys import PublicKey, GPGKey, DeployKey
 from .secrets import UserSecret, OrgSecret, RepoSecret
 from .tasks import Action, SpecKind, Task
 from .works import Work
+from .doe_spec import DoeSpec
+from .evolve_spec import EvolveSpec
+from .project_payload import ProjectPayload
+from .peagen_toml_spec import PeagenTomlSpec
+from .eval_result import EvalResult
+from .analysis_result import AnalysisResult
+from .pools import Pool
+from .workers import Worker
 from .mixins import RepositoryMixin, RepositoryRefMixin
-from peagen.defaults import (
-    DEFAULT_TENANT_ID,
-    DEFAULT_TENANT_EMAIL,
-    DEFAULT_TENANT_NAME,
-    DEFAULT_TENANT_SLUG,
-)
-
-# ---------------------------------------------------------------------
 
 
 def _is_terminal(cls, state: str | Status) -> bool:
@@ -59,102 +38,6 @@ def _is_terminal(cls, state: str | Status) -> bool:
 
 
 Status.is_terminal = classmethod(_is_terminal)
-
-
-# ---------------------------------------------------------------------
-# Repository hierarchy
-# ---------------------------------------------------------------------
-class Tenant(TenantBase, Bootstrappable):
-    DEFAULT_ROWS = [
-        {
-            "id": DEFAULT_TENANT_ID,
-            "email": DEFAULT_TENANT_EMAIL,
-            "name": DEFAULT_TENANT_NAME,
-            "slug": DEFAULT_TENANT_SLUG,
-        }
-    ]
-
-
-class User(UserBase, Bootstrappable):
-    pass
-    # DEFAULT_ROWS = [
-    #     {
-    #         "id": DEFAULT_SUPER_USER_ID,
-    #         "email": DEFAULT_SUPER_USER_EMAIL,
-    #         "tenant_id": DEFAULT_TENANT_ID,
-    #     },
-    #     {
-    #         "id": DEFAULT_SUPER_USER_ID_2,
-    #         "email": DEFAULT_SUPER_USER_EMAIL_2,
-    #         "tenant_id": DEFAULT_TENANT_ID,
-    #     }
-    # ]
-
-
-class Repository(Base, GUIDPk, Timestamped, TenantBound, Ownable, StatusMixin):
-    """
-    A code or data repository that lives under a tenant.
-    – parent of Secrets & DeployKeys
-    """
-
-    __tablename__ = "repositories"
-    __table_args__ = (
-        UniqueConstraint("url"),
-        UniqueConstraint("tenant_id", "name"),
-    )
-    name = Column(String, nullable=False)
-    url = Column(String, unique=True, nullable=False)
-    default_branch = Column(String, default="main")
-    commit_sha = Column(String(length=40), nullable=True)
-    remote_name = Column(String, nullable=False, default="origin")
-
-    secrets = relationship(
-        "RepoSecret", back_populates="repository", cascade="all, delete-orphan"
-    )
-    deploy_keys = relationship(
-        "DeployKey", back_populates="repository", cascade="all, delete-orphan"
-    )
-    tasks = relationship(
-        "Task",
-        back_populates="repository",
-        cascade="all, delete-orphan",
-    )
-
-
-# ---------------------------------------------------------------------
-# association edges
-# ---------------------------------------------------------------------
-
-
-# class UserTenant(Base, GUIDPk, TenantMixin, UserMixin):
-#     """
-#     Many-to-many edge between users and tenants.
-#     A user may be invited to / removed from any number of tenants.
-#     """
-
-#     __tablename__ = "user_tenants"
-#     joined_at = Column(tzutcnow, default=dt.timezone.utcnow, nullable=False)
-
-
-class UserRepository(Base, GUIDPk, RepositoryMixin, UserMixin):
-    """
-    Edge capturing *any* per-repository permission or ownership
-    the user may have.  `perm` can be "owner", "push", "pull", etc.
-    """
-
-    __tablename__ = "user_repositories"
-
-
-# ---------------------------------------------------------------------
-# 5) Raw blobs (stand-alone)
-# ---------------------------------------------------------------------
-
-
-class RawBlob(Base, GUIDPk, Timestamped, BlobRef):
-    __tablename__ = "raw_blobs"
-    mime_type = Column(String, nullable=False)
-    size = Column(Integer, nullable=False)
-
 
 __all__ = [
     "Tenant",
@@ -182,4 +65,10 @@ __all__ = [
     "Task",
     "Work",
     "RawBlob",
+    "DoeSpec",
+    "EvolveSpec",
+    "ProjectPayload",
+    "PeagenTomlSpec",
+    "EvalResult",
+    "AnalysisResult",
 ]
